Add ClientRepository, ClientService, AccountRepository, AccountService

+++ src/main/java/com/example/clertecbankproject/model/repository/impl/BankRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.example.clertecbankproject.model.repository.impl;
 
 import com.example.clertecbankproject.model.bd.postgresql.JDBCPostgreSQLConnection;
 import com.example.clertecbankproject.model.entity.Bank;
+import com.example.clertecbankproject.model.entity.Client;
 import com.example.clertecbankproject.model.repository.BankRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -40,15 +41,21 @@ public class BankRepositoryImpl implements BankRepository {
                 banks.add(bank);
             }
         } catch (SQLException e) {
-            System.out.println(e.getMessage());
+            logger.error(e.getMessage());
         }
         return banks;
     }
 
     @Override
     public void deleteBank(Long id) throws Exception {
-        String deleteTableSQL = "DELETE FROM banks WHERE id = " + id;
-        executeStatement(deleteTableSQL);
+        Bank bank = getBank(id);
+        if(bank.getId()!=null){
+            String deleteTableSQL = "DELETE FROM banks WHERE id = " + id;
+            executeStatement(deleteTableSQL);
+            logger.info("Bank with id= '{}' delete", id);
+        }else {
+            logger.debug("Bank with id={} don't exist",id);
+        }
     }
 
     @Override
@@ -79,6 +86,46 @@ public class BankRepositoryImpl implements BankRepository {
         String updateTableSQL = "UPDATE banks SET bank_name ="
                 + "'" + nameOfBank + "'" + "WHERE id = " + idForUpdate;
         executeStatement(updateTableSQL);
+    }
+
+    @Override
+    public void addClientToBank( Long bankId, Long clientId) {
+        String insertTableSQL = "INSERT INTO bank_clients"
+                + "(bank_id, client_id) " + "VALUES "
+                + "(" + bankId + ", "
+                + clientId + ")";
+        executeStatement(insertTableSQL);
+
+    }
+
+    @Override
+    public List<Client> showAllBankClients(Long bankId) throws Exception{
+        Connection connection = null;
+        List<Client> clients = new ArrayList<>();
+        try {
+            logger.info("Connecting to a database...");
+            connection = JDBCPostgreSQLConnection.getConnection();
+            logger.info("Connected database successfully...");
+            String selectTableSQL = "SELECT c.id, c.first_name, c.second_name, c.last_name, c.address " +
+                    "from clients c " +
+                    "left join bank_clients bc " +
+                    "on c.id = bc.client_id " +
+                    "where bc.bank_id = " + bankId;
+            PreparedStatement pstmt = connection.prepareStatement(selectTableSQL);
+            ResultSet rs = pstmt.executeQuery();
+            while (rs.next()) {
+                Long id = Long.parseLong(rs.getString("id"));
+                String firstName = rs.getString("first_name");
+                String secondName = rs.getString("second_name");
+                String lastName = rs.getString("last_name");
+                String address = rs.getString("address");
+                Client client = new Client(id, firstName, secondName, lastName, address);
+                clients.add(client);
+            }
+        } catch (SQLException e) {
+            logger.error(e.getMessage());
+        }
+        return clients;
     }
